+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 All notable changes to the "flutter-monorepo-utils" extension will be documented in this file.
 
 ## [0.0.2] - 13/07/2021
-- Adds a validation to the check the selected folder before execute the test commands
+
+- Adds a validation to check the selected folder before execute the test commands
 - Improve Test runner documentation
+- Fix Test runner execution on Windows
 
 ## [0.0.1] - 13/07/2021